Clean up remaining build warnings
Clean up remaining build warnings.  None of this should affect code
operations.  The only warnings which were actually relevant might have
been the ones in ldrdf.c, but it's not clear if anyone ever uses that.

+++ output/outobj.c
@@ -1995,7 +1995,7 @@ static void obj_write_file(int debuginfo)
     orp->parm[0] = 0;
     orp->parm[1] = 0;
     for (pub = fpubhead; pub; pub = pub->next) {        /* pub-crawl :-) */
-        if ((uint64_t)orp->parm[2] != pub->segment) {
+        if (orp->parm[2] != (uint32_t)pub->segment) {
             obj_emit(orp);
             orp->parm[2] = pub->segment;
         }
